+++ Content.Server/AU14/Round/AuVoteRuleComponent.cs
@@ -1,0 +1,9 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Server.AU14.Round;
+
+[RegisterComponent]
+public sealed partial class AuVoteRuleComponent : Component
+{
+
+}
