--- Content.Server/AU14/Round/PlanetPlatoonVoteRuleComponent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Server.AU14.Round;
-
-[RegisterComponent]
-public sealed partial class PlanetPlatoonVoteRuleComponent : Component
-{
-
-}
