+++ Content.Server/AU14/VendorMarker/VendorMarkerComponent.cs
@@ -1,4 +1,5 @@
 using Content.Shared.AU14;
+using Content.Shared.AU14.util;
 using Robust.Shared.GameObjects;
 
 namespace Content.Server.AU14.VendorMarker
