+++ Content.Server/CharacterInfo/CharacterInfoSystem.cs
@@ -1,9 +1,9 @@
 ﻿using Content.Server.Mind;
 using Content.Server.Roles;
 using Content.Server.Roles.Jobs;
+using Content.Shared.Au14.Util;
 using Content.Shared.CharacterInfo;
 using Content.Shared.Inventory;
-using Content.Shared.JobTitleChanger;
 using Content.Shared.Objectives;
 using Content.Shared.Objectives.Components;
 using Content.Shared.Objectives.Systems;
