--- Content.Shared/AU14/ColonyCommsConsoleComponent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared.AU14;
-
-[RegisterComponent]
-public sealed partial class ColonyCommsConsoleComponent : Component
-{
-    public bool SirenActive = false;
-}
