+++ Content.Shared/AU14/ColonyEconomy/ColonyCommsConsoleComponent.cs
@@ -1,0 +1,9 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.AU14;
+
+[RegisterComponent]
+public sealed partial class ColonyCommsConsoleComponent : Component
+{
+    public bool SirenActive = false;
+}
