+++ Content.Shared/AU14/ColonyEconomy/ColonySirenComponent.cs
@@ -1,0 +1,12 @@
+// Content.Server/AU14/ColonyEconomy/BudgetConsoleComponent.cs
+using Robust.Shared.GameStates;
+using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
+
+namespace Content.Shared.AU14;
+
+[RegisterComponent]
+public sealed partial class ColonySirenComponent : Component
+{
+
+
+}
