--- Content.Shared/AU14/ColonySirenComponent.cs
@@ -1,12 +1,0 @@
-// Content.Server/AU14/ColonyEconomy/BudgetConsoleComponent.cs
-using Robust.Shared.GameStates;
-using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
-
-namespace Content.Shared.AU14;
-
-[RegisterComponent]
-public sealed partial class ColonySirenComponent : Component
-{
-
-
-}
