--- Content.Shared/AU14/GamePlanetPoolPrototype.cs
@@ -1,17 +1,0 @@
-using Robust.Shared.Prototypes;
-using System.Collections.Generic;
-using JetBrains.Annotations;
-
-namespace Content.Shared.AU14
-{
-    [Prototype("GamePlanetPool"),PublicAPI]
-    public sealed partial class GamePlanetPoolPrototype : IPrototype
-    {
-        [IdDataField]
-        public string ID { get; private set; } = string.Empty;
-
-        [DataField("planets")]
-        public List<string> Planets { get; private set; } = new();
-    }
-}
-
