+++ Content.Shared/AU14/Util/PlatoonJobClass.cs
@@ -1,0 +1,16 @@
+namespace Content.Shared.AU14.util
+{
+    public enum PlatoonJobClass
+    {
+        SquadAutomaticRifleman,
+        PlatoonCorpsman,
+        PlatoonCommander,
+        RadioTelephoneOperator,
+        SectionSergeant,
+        SquadRifleman,
+        Pilot,
+        OperationsOfficer,
+
+    }
+}
+
