+++ Content.Shared/AU14/Util/PlatoonMarkerClass.cs
@@ -1,0 +1,27 @@
+namespace Content.Shared.AU14.util
+{
+    public enum PlatoonMarkerClass
+    {
+        Corpsman,
+        Clothing,
+        Weapons,
+        SWeapons,
+        Arifleman,
+        Rifleman,
+        Dcc,
+        OperationsOfficer,
+        Rto,
+        SectionSergeant,
+        Pilot,
+        ExtraVendor1,
+        ExtraVendor2,
+        ExtraVendor3,
+        ExtraVendor4,
+        Deco1,
+        Deco2,
+        Deco3,
+        Deco4,
+        Deco5,
+        Deco6
+    }
+}
