+++ Content.Shared/_RMC14/Rules/RMCPlanetMapPrototypeComponent.cs
@@ -1,5 +1,6 @@
 using Content.Shared._RMC14.Item;
 using Content.Shared.AU14;
+using Content.Shared.AU14.util;
 using Content.Shared.Roles;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
