--- Resources/Prototypes/_AU14/Entities/crates.yml
@@ -1,119 +1,0 @@
-- type: entity
-  parent: RMCCrateBase
-  id: AU14CrateCivilianClothingRandom
-  name: Crate of Random Civilian Clothing
-  description: Contains a random assortment of civilian clothing items, where did they come from?
-  components:
-  - type: StorageFill
-    contents:
-    - id: AU14CrateCivilianClothingRandomSelection
-      amount: 15
-
-
-# in Resources/Prototypes/_AU14/Entities/crates.yml
-
-- type: entity
-  id: AU14CrateCivilianClothingRandomSelection
-  name: Random Civilian Clothing
-  suffix: Random
-  components:
-  - type: RandomSpawner
-    prototypes:
-    - RMCShoesBlack
-    - RMCShoesBlue
-    - RMCShoesBrown
-    - RMCShoesGreen
-    - RMCShoesOrange
-    - RMCShoesPurple
-    - RMCShoesRed
-    - RMCShoesWhite
-    - RMCShoesYellow
-    - RMCBootsPMC
-    - RMCBootsVanBandolier
-    - RMCShoesLaceupBrown
-    - RMCShoesLaceup
-    - RMCShoesLeather
-    - AU14CivilianWorkwearYellow
-    - AU14CivilianWorkwearPink
-    - AU14CivilianWorkwearBlue
-    - AU14CivilianWorkwearGreen
-    - CMJumpsuitLiaisonBlack
-    - CMJumpsuitLiaisonBlue
-    - CMJumpsuitLiaisonBrown
-    - CMJumpsuitLiaisonField
-    - CMJumpsuitLiaisonIvy
-    - CMJumpsuitLiaisonCorporateFormal
-    - CMJumpsuitLiaisonFormal
-    - CMJumpsuitLiaison
-    - CMJumpsuitLiaisonBlazer
-    - CMJumpsuitLiaisonCharcoal
-    - RMCJumpsuitDutchBandolier
-    - RMCJumpsuitCivilian
-    - CMJumpsuitColonist
-    - CMJumpsuitTShirtWhite
-    - CMJumpsuitTShirtGray
-    - CMJumpsuitTShirtRed
-    - RMCJacketCorporateKhaki
-    - RMCJacketCorporateBlack
-    - RMCJacketCorporateBlue
-    - RMCJacketCorporateBrown
-    - AU14CivilianJacketGrayPufferVest
-    - AU14CivilianJacketKhakiPufferVest
-    - AU14CivilianJacketTanPufferVest
-    - AU14CivilianJacketGrayPufferJacket
-    - AU14CivilianJacketOrangePufferJacket
-    - AU14CivilianJacketKhakiPufferJacket
-    - AU14WindbreakerBlue
-    - AU14WindbreakerGray
-    - AU14WindbreakerGreen
-    - AU14WindbreakerKhakiGreenMix
-    - AU14WindbreakerKhaki
-    - AU14CivilianJacketBlueParka
-    - AU14CivilianJacketGreenParka
-    - AU14CivilianJacketRedParka
-    - AU14CivilianJacketYellowParka
-    - AU14CivilianTanTrenchCoat
-    - AU14CivilianBrownTrenchCoat
-    - AU14CivilianGrayTrenchCoat
-    - AU14CivilianJacketBomberJacket
-    - AU14CivilianJacketOldCoat
-    - AU14CivilianJacketSnowSuit
-    - AU14CivBallCapBlack
-    - AU14CivBallCapBlueTrucker
-    - AU14CivBallCapRedTrucker
-    - AU14USCMVeteranCap
-    - AU14CivFedoraTan
-    - AU14CivFedoraBrown
-    - AU14CivFedoraGray
-    - RMCSunglasses
-    - RMCHipsterGlasses
-    - RMCSunglassesBig
-    - AU14GlassesPersonalOrange
-    - RMCGlassesAviators
-    - RMCGlassesTriMaxYellow
-    - RMCGlassesTriMaxBlack
-    - RMCGlassesTriMaxBronze
-    - AU14GlassesBiMexOrange
-    - CMHandsBrown
-    - CMHandsLightBrown
-    - CMSatchel
-    - AU14SlingSatchelBlack
-    - AU14SlingSatchelBlue
-    - RMCBeltUtilityGeneral
-    - RMCPouchGeneral
-
-
-- type: entity
-  parent: RMCCrateBase
-  id: AU14CrateWorkTools
-  name: Crate of Work Tools
-  description: Contains several fire axes and pickaxes for heavy-duty work.
-  components:
-  - type: Sprite
-    state: crate
-  - type: StorageFill
-    contents:
-    - id: RMCFireAxe
-      amount: 3
-    - id: RMCPickaxe
-      amount: 3
