+++ Resources/Prototypes/_AU14/Maps/lv624.yml
@@ -1,0 +1,46 @@
+- type: gameMap
+  id: LV624
+  mapName: LV-624
+  mapPath: /Maps/_AU14/LV624.yml
+  minPlayers: 0
+  maxPlayers: 75
+  maxRandomOffset: 0
+  randomRotation: false
+  stations:
+    lv624colonyfallcompmap:
+      stationProto: StandardUNMCWarship
+      components:
+        - type: StationNameSetup
+          mapNameTemplate: 'lv624colonyfallcompmap'
+        - type: StationJobs
+          availableJobs:
+            AU14JobCivilianColonist: [ -1, -1 ]
+            AU14JobCivilianFoodServiceWorker: [2, 2]
+            AU14JobCivilianJournalist: [1, 1]
+            AU14JobCivilianScientist: [3, 3]
+            AU14JobCivilianColonySynthetic: [1, 1]
+
+            # Colony Command (non department)
+            AU14JobCivilianCorporateLiaison: [1, 1]
+            AU14JobCivilianColonyAdministrator: [1, 1]
+
+            # Security/LEO
+            AU14JobCivilianCMBMarshal: [1, 1]
+            AU14JobCivilianCMBDeputy: [1, 1]
+
+            # Engineering/Logi
+            AU14JobCivilianForeman: [1, 1]
+            AU14JobCivilianFreightSystemsSpecialist: [1, 1]
+            AU14JobCivilianEngineer: [3, 3]
+            AU14JobCivilianWasteManagementSpecialist: [2, 2]
+
+            # Medical
+            AU14JobCivilianPhysician: [1, 1]
+            AU14JobCivilianNurse: [1, 1]
+            AU14JobCivilianEthicsAndWellnessAdvisor: [1, 1]
+            AU14JobCivilianEmergencyResponseOfficer: [1, 1]
+
+            # Other
+            AU14JobCivilianUSASFRecruiter: [1, 1]
+
+
