+++ Resources/Prototypes/_AU14/gamemode/platoons.yml
@@ -1,0 +1,79 @@
+- type: platoon
+  id: LACN
+  name: LACN Platoon
+  VendorToMarker:
+    Rifleman: AU14LACNequipmentvendor
+    Clothing: AU14LACNclothingequipmentvendor
+    Weapons: AU14LACNWeaponsVendor
+    SWeapons: AU14LACNSpecialWeaponsVendor
+    Corpsman: AU14LACNcorpsmanvendor
+    Arifleman: AU14LACNMGequipmentvendor
+    Dcc: AU14LACNaircrewvendor
+    OperationsOfficer: AU14LACNcommandequipmentvendor
+    Rto: AU14RTOVendor
+    SectionSergeant: AU14USCMsectionleaderequipmentvendor
+    Pilot: AU14AircrewWeaponsVendor
+  possibleships:
+    - Riptide
+
+
+- type: platoon
+  id: USCM
+  name: USCM Platoon
+  VendorToMarker:
+    Rifleman: AU14USCMequipmentvendor
+    Clothing: AU14USCMclothingequipmentvendor
+    Weapons: AU14USCMWeaponsVendor
+    SWeapons: AU14USCMSpecialWeaponsVendor
+    Corpsman: AU14USCMcorpsmanvendor
+    Arifleman: AU14USCMSGequipmentvendor
+    Dcc: AU14USCMaircrewvendor
+    OperationsOfficer: AU14USCMcommandequipmentvendor
+    Rto: AU14RTOVendor
+    SectionSergeant: AU14USCMsectionleaderequipmentvendor
+    Pilot: AU14AircrewWeaponsVendor
+    ExtraVendor1: VMarkerOpforDeco5
+    Deco1: VMarkerOpforDeco5
+  possibleships:
+      - Riptide
+
+- type: platoon
+  id: UPP
+  name: UPP Platoon
+  VendorToMarker:
+    Rifleman: AU14USCMequipmentvendor
+    Corpsman: AU14USCMcorpsmanvendor
+    Arifleman: AU14USCMequipmentvendor
+    Dcc: AU14AircrewWeaponsVendor
+    OperationsOfficer: AU14USCMsectionleaderequipmentvendor
+    Rto: AU14RTOVendor
+    SectionSergeant: AU14USCMsectionleaderequipmentvendor
+    Pilot: AU14AircrewWeaponsVendor
+  possibleships:
+    - Riptide
+  jobClassOverride:
+    SquadRifleman: RMCSPPRifleman
+    SquadAutomaticRifleman: AU14JobCivilianColonyAdministrator
+  jobSlotOverrideGovfor:
+    AU14JobGOVFORSquadRifleman: 16
+    AU14JobGOVFORPlatoonCorpsman: 2
+    AU14JobGOVFORPlatCo: 1
+  jobSlotOverrideOpfor:
+    AU14JobOpforSquadRifleman: 12
+    AU14JobOpforPlatoonCorpsman: 1
+    AU14JobOpforPlatCo: 1
+
+
+- type: platoon
+  id: WEYU
+  name: Weyland-Yutani Platoon
+  VendorToMarker:
+    Rifleman: AU14USCMequipmentvendor
+    Corpsman: AU14USCMcorpsmanvendor
+    Arifleman: AU14USCMequipmentvendor
+    Dcc: AU14AircrewWeaponsVendor
+    OperationsOfficer: AU14USCMsectionleaderequipmentvendor
+    Rto: AU14RTOVendor
+    SectionSergeant: AU14USCMsectionleaderequipmentvendor
+    Pilot: AU14AircrewWeaponsVendor
+# Add more platoons as needed
