--- Resources/Prototypes/_AU14/gamemodes.yml
@@ -1,305 +1,0 @@
-- type: gamePreset
-  id: ForceOnForce
-  alias:
-    - forceonforce
-  name: Force on Force
-  showInVote: true
-  description: Force on Force mode
-  requiresOpforVote: true
-  requiresGovforVote: true
-  supportedPlanets:
-    - AUPlanetLV747
-    - AUPlanetShepherdsPride
-  rules:
-  - PlanetPlatoonVoteRule
-  - PlatoonSpawn
-  - AddGovfor
-  - AddOpfor
-
-- type: gamePreset
-  id: Insurgency
-  alias:
-    - insurgency
-  name: Insurgency
-  showInVote: true
-  description: Insurgency mode
-  requiresGovforVote: true
-  supportedPlanets:
-    - AUPlanetLV747
-    - AUPlanetShepherdsPride
-    - AUPlanetLV624
-  rules:
-  - PlanetPlatoonVoteRule
-  - PlatoonSpawn
-  - AddGovfor
-  - AddClf
-  - ColonyAntags
-
-- type: gamePreset
-  id: ColonyFall
-  alias:
-    - colonyfall
-  name: Colony Fall
-  showInVote: true
-  description: Colony Fall mode
-  rules:
-    - PlanetPlatoonVoteRule
-    - ColonyAntags
-  supportedPlanets:
-    - AUPlanetLV747
-    - AUPlanetShepherdsPride
-    - AUPlanetLV624
-
-- type: gamePreset
-  id: Jailbreak
-  alias:
-    - jailbreak
-  name: Jailbreak
-  showInVote: false
-  description: Jailbreak mode
-  rules:
-    - PlanetPlatoonVoteRule
-
-- type: gamePreset
-  id: Prometheus
-  alias:
-    - prometheus
-  name: Prometheus
-  showInVote: false
-  description: Prometheus mode
-  rules:
-    - PlanetPlatoonVoteRule
-
-- type: gamePreset
-  id: Criminal
-  alias:
-    - criminal
-  name: Criminal
-  showInVote: false
-  description: Criminal mode
-  rules:
-    - PlanetPlatoonVoteRule
-
-
-
-
-
-
-
-- type: entity
-  id: PlatoonSpawn
-  parent: BaseGameRule
-  components:
-  - type: PlatoonSpawnRule
-
-- type: entity
-  id: AddGovfor
-  parent: BaseGameRule
-  components:
-  - type: AddJobsRule
-    addToShip: true
-    shipFaction: Govfor
-    jobs:
-      AU14JobGOVFORPlatCo: 1
-      AU14JobGOVFORPlatoonCorpsman: 3
-      AU14JobGOVFORDCC: 2
-      AU14JobGOVFORPlatOp: 1
-      AU14JobGOVFORDSPilot: 1
-      AU14JobGOVFORRadioTelephoneOperator: 1
-      AU14JobGOVFORSectionSergeant: 1
-      AU14JobGOVFORSquadAutomaticRifleman: 1
-      AU14JobGOVFORSquadRifleman: 14
-
-
-- type: entity
-  id: AddOpfor
-  parent: BaseGameRule
-  components:
-  - type: AddJobsRule
-    jobs:
-      AU14JobOpforPlatCo: 1
-      AU14JobOpforPlatoonCorpsman: 3
-      AU14JobOpforDCC: 2
-      AU14JobOpforPlatOp: 1
-      AU14JobOpforDSPilot: 1
-      AU14JobOpforRadioTelephoneOperator: 1
-      AU14JobOpforSectionSergeant: 1
-      AU14JobOpforSquadAutomaticRifleman: 1
-      AU14JobOpforSquadRifleman: 14
-
-- type: entity
-  id: AddClf
-  parent: BaseGameRule
-  components:
-  - type: AddJobsRule
-    jobs:
-      AU14JobCLFCellLeader: 1
-      AU14JobCLFPhysician: 1
-      AU14JobCLFGuerilla: 6
-      AU14JobCLFMachineGunner: 1
-
-- type: entity
-  id: rockwallremoveinssolaris
-  name: Rock Wall Remove Solaris (Insurgency)
-  parent: CMWallSolarisRock
-  components:
-    - type: DestroyOnPreset
-      preset: Insurgency
-
-
-- type: entity
-  id: rockwallremoveinshybrisia
-  name: Rock Wall Remove hybrisia (Insurgency)
-  parent: RMCWallHybrisaRock
-  components:
-  - type: DestroyOnPreset
-    preset: Insurgency
-
-
-
-- type: entity
-  id: rockwallremoveinshbyrisia
-  name: Rock Wall Remove Kutjevo (Insurgency)
-  parent: RMCWallKutjevoRock
-  components:
-  - type: DestroyOnPreset
-    preset: Insurgency
-
-- type: entity
-  id: rockwallremovein
-  name: Rock Wall Remove  (Insurgency)
-  parent: CMWallRock
-  components:
-  - type: DestroyOnPreset
-    preset: Insurgency
-
-- type: entity
-  id: barrierremove
-  name: Invis Barrier Remove (Insurgency)
-  parent: RMCBarrierBase
-  components:
-  - type: DestroyOnPreset
-    preset: Insurgency
-
-- type: entity
-  id: rockwallremoveinssolaris_forceonforce
-  name: Rock Wall Remove Solaris (ForceOnForce)
-  parent: CMWallSolarisRock
-  components:
-    - type: DestroyOnPreset
-      preset: ForceOnForce
-
-- type: entity
-  id: rockwallremoveinshybrisia_forceonforce
-  name: Rock Wall Remove Hybrisia (ForceOnForce)
-  parent: RMCWallHybrisaRock
-  components:
-    - type: DestroyOnPreset
-      preset: ForceOnForce
-
-- type: entity
-  id: rockwallremoveinshbyrisia_forceonforce
-  name: Rock Wall Remove Kutjevo (ForceOnForce)
-  parent: RMCWallKutjevoRock
-  components:
-    - type: DestroyOnPreset
-      preset: ForceOnForce
-
-- type: entity
-  id: rockwallremoveinssolaris_criminal
-  name: Rock Wall Remove Solaris (Criminal)
-  parent: CMWallSolarisRock
-  components:
-    - type: DestroyOnPreset
-      preset: Criminal
-
-- type: entity
-  id: rockwallremoveinshybrisia_criminal
-  name: Rock Wall Remove Hybrisia (Criminal)
-  parent: RMCWallHybrisaRock
-  components:
-    - type: DestroyOnPreset
-      preset: Criminal
-
-- type: entity
-  id: rockwallremoveinshbyrisia_criminal
-  name: Rock Wall Remove Kutjevo (Criminal)
-  parent: RMCWallKutjevoRock
-  components:
-    - type: DestroyOnPreset
-      preset: Criminal
-
-- type: entity
-  id: rockwallremoveinssolaris_prometheus
-  name: Rock Wall Remove Solaris (Prometheus)
-  parent: CMWallSolarisRock
-  components:
-    - type: DestroyOnPreset
-      preset: Prometheus
-
-- type: entity
-  id: rockwallremoveinshybrisia_prometheus
-  name: Rock Wall Remove Hybrisia (Prometheus)
-  parent: RMCWallHybrisaRock
-  components:
-    - type: DestroyOnPreset
-      preset: Prometheus
-
-- type: entity
-  id: rockwallremoveinshbyrisia_prometheus
-  name: Rock Wall Remove Kutjevo (Prometheus)
-  parent: RMCWallKutjevoRock
-  components:
-    - type: DestroyOnPreset
-      preset: Prometheus
-
-- type: entity
-  id: rockwallremoveinssolaris_colonyfall
-  name: Rock Wall Remove Solaris (ColonyFall)
-  parent: CMWallSolarisRock
-  components:
-    - type: DestroyOnPreset
-      preset: ColonyFall
-
-- type: entity
-  id: rockwallremoveinshybrisia_colonyfall
-  name: Rock Wall Remove Hybrisia (ColonyFall)
-  parent: RMCWallHybrisaRock
-  components:
-    - type: DestroyOnPreset
-      preset: ColonyFall
-
-- type: entity
-  id: rockwallremoveinshbyrisia_colonyfall
-  name: Rock Wall Remove Kutjevo (ColonyFall)
-  parent: RMCWallKutjevoRock
-  components:
-    - type: DestroyOnPreset
-      preset: ColonyFall
-
-- type: entity
-  id: ColonyAntags
-  parent: BaseGameRule
-  name: Colony Antags
-  description: Randomly rolls for each colony antag to be present this round.
-  components:
-  - type: GameRule
-  - type: ColonyAntagsRule
-
-
-- type: entity
-  id: PlanetPlatoonVoteRule
-  parent: BaseGameRule
-  name: Colony Antags
-  description: If you want your mode to load planets and platoons automatically
-  components:
-  - type: GameRule
-  - type: PlanetPlatoonVoteRule
-
-- type: GamePlanetPool
-  id: ColonyPlanetPool
-  planets:
-    - AUPlanetLV747
-    - AUPlanetShepherdsPride
-    - AUPlanetStableGarrison
-    - AUPlanetLV624
